% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazardcross.R
\docType{data}
\name{hazardcross}
\alias{hazardcross}
\title{Simulated Survival with Crossing Hazard Functions}
\format{The \code{hazardcross} is a data frame with 100 simulated observations of 3 variables,
and has the following columns:
\itemize{
   \item \code{time} the survival time
   \item \code{censor} the censoring indicator
   \item \code{group} the grouping variable
}}
\usage{
hazardcross
}
\description{
The data frame \code{hazardcross} is simulated from two groups of piecewise exponential
lifetime distributions with crossing hazard functions. The estimated survival functions remain ordered even when the estimated hazard 
functions are crossed.
See \code{\link{supELtest}} for the application.
}
\seealso{
\code{\link{supELtest}}
}
\keyword{datasets}
